Major changes to Train.html and FunctionSpec.js to implement the new training trial sampler (N.B. this job is incomplete, see Note 1 below). These changes also included the creation of 2 new JS files (GlobalVars.js and ObjectSpec.js) that better organise the existing code. A PromiseChain() function was declared in FunctionSpec.js (called in Train.html) to synchronise the execution of dependent functions: GetTrainHist(), SeedRng(), RunTrailLoop().
Minor changes to GetTrainHist.php to update the pair count to 49 since the second operation has been removed.

Additional changes to ExclusionCheck.js blocking the use of MS Edge.

Some unnecessary files/directories deleted. Multiple pre-existing but untracked files/directories added to the repository.

Note 1: The function GetTimelineVars() that generates the TimelineVars array needs to be re-written so that it returns an array that is properly ordered by PairId (i.e., not having all the Supervised Pairs first, and the Unsupervised pairs second). Currently, the training history array (TrainHistory) as returned by GetTrainHistory.php knows nothing about order of trails in the TimelineVars array so there is a mismatch between the actual and target potability distributions when sampling pairs. Additionally, we should probably rename the TimelineVars array to something more suitable, reflecting the fact that we are no longer getting jsPscyh to control the order of trials.

On branch master
Changes to be committed:
	new file:   Assets/jquery.min.js
	new file:   Assets/jquery.min.map
	new file:   Assets/ua-parser.pack.js
	new file:   Documents/ConsentForm_MRI.pdf
	new file:   Documents/InfoSheetConsentForm_Web.pdf
	new file:   Documents/InfoSheet_MRI.pdf
	new file:   Documents/SONA_advert.pdf
	modified:   ExclusionCheck.js
	modified:   FunctionSpec.js
	modified:   GetTrainHist.php
	new file:   GlobalVars.js
	new file:   Imgs/S0.png
	new file:   Imgs/S1.png
	new file:   Logo.png
	new file:   ObjectSpec.js
	new file:   Symbols/olds00.png
	new file:   Symbols/olds01.png
	new file:   Symbols/s00.png
	new file:   Symbols/s01.png
	new file:   Symbols/symbol 1.png
	new file:   TestShapes/Basic_Image_presentations.html
	new file:   TestShapes/TestShowImgs.html
	new file:   TestShapes/i00.jpg
	new file:   TestShapes/i01.jpg
	new file:   TestShapes/i02.jpg
	new file:   TestShapes/i03.jpg
	new file:   TestShapes/i04.jpg
	new file:   TestShapes/i05.jpg
	new file:   TestShapes/i06.jpg
	modified:   Train.html
	new file:   Trash/Backup_Question_v03.html
	new file:   Trash/GetArrayOfResponseTags.js
	new file:   Trash/Question_20231127-1024.html
	new file:   Trash/Question_20231127-1400.html
	new file:   Trash/Question_v01.html
	new file:   Trash/Question_v02.html
	new file:   Trash/Question_v03.html
	new file:   Trash/ReadTaskIO.php
	new file:   index.html