Bug fix for GetLearnStats.php ... ... in response to the following PHP error:     PHP Fatal error:  Uncaught TypeError: array_push(): Argument #1 ($array) must be of type array, null given in /home/learning/b01.learningandinference.org/GetLearnStats.php:104
This error was cause by failure to log the first attempt in a session (probably due error transmission errors).
Unrecorded first attempts meant that while loop starting on line 79 was would initially see a $AttemptNum greater than zero.
This invokes the elseif statement on 103, before the trial counter ($iT) has been incremented passed -1 (its initial state).
As a result, the following code attempted to push objects in $Trials[-1] which did not exist.
The fix provided in this commit detects when the first attempt in a session does not have an $AttemptNum equal 0 and fills in the missing data.